django-backend/requirements.txt
# Core Django
Django==4.2.20
djangorestframework==3.14.0
django-cors-headers==4.7.0
django-filter==23.5
python-dotenv==1.0.1

# Authentication
PyJWT==2.8.0
djangorestframework-simplejwt==5.3.1
google-auth-oauthlib==1.2.0
google-auth==2.27.0
python-jose==3.3.0
passlib==1.7.4

# Database
psycopg2-binary==2.9.9
pymongo==4.6.1
motor==3.3.2

# Utilities
requests==2.32.3
python-dateutil==2.8.2
beautifulsoup4==4.13.3

# AI/ML (existing)
langchain==0.3.24
langchain_community==0.3.23

# Development
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
coverage==7.3.2
